+++ nexus/usecases/agents/agents.py
@@ -51,6 +51,7 @@ class AgentDTO:
     foundation_model: str = None
     guardrail_configuration: dict = None
     credentials: List[Dict] = None
+    source_type: str = None
 
     def dict(self):
         return {key: value for key, value in self.__dict__.items() if value is not None}
@@ -166,6 +167,7 @@ class AgentUsecase:
             tags=agent_dto.tags,
             model_id=agent_dto.model[0],
         )
+
         return external_id
 
     def create_contact_fields(self, project_uuid: str, fields: List[Dict[str, str]], agent, convert_fields: bool = True):
@@ -713,6 +715,7 @@ class AgentUsecase:
             foundation_model=agent_value.get("foundation_model"),
             guardrail_configuration=agent_value.get("guardrail_configuration"),
             credentials=agent_value.get("credentials"),
+            source_type=agent_value.get("source_type"),
         )
         validate_agents = self.validate_agent_dto(agent_dto, user_email)
         return validate_agents